fix a bug in crossbar that causes response buf not being released

+++ src/crossbar/tlm_gp_mm.h
@@ -4,7 +4,7 @@
 #include <tlm>
 #include <queue>
 
-typedef uint64_t trans_id_t;
+typedef unsigned long long trans_id_t;
 
 class tlm_gp_mm : public tlm::tlm_mm_interface
 {
